pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlbot-pro"
version = "0.1.0"
description = "Production-ready Rocket League bot foundation"
authors = [{name = "RLBot Pro Team"}]
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.4",
    "pydantic==2.7.0",
    "PyYAML==6.0.1",
    "matplotlib==3.8.4",
    "pytest==7.4.4",
    "mypy==1.8.0",
    "ruff==0.3.5",
    "black==24.3.0",
]
readme = "README.md"
license = {file = "LICENSE"}

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "S", "BLE", "C4", "DTZ", "EM", "RET", "TID"]
ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
implicit_reexport = false
mypy_path = ["."]

[tool.pytest.ini_options]
addopts = "-q --disable-warnings"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["rlbot_pro*", "telemetry*", "reports*", "scenarios*", "gui*"]
